+++ CHANGELOG.md
@@ -35,3 +35,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch publications and other documents
 - Documents Table
 - project-code as parameter in project-view
+- Bridgehead state