+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-02-26]
+## [1.1.0 - 2024-02-27]
 ### Added
 - Dockerfile
 - Single Spa
@@ -37,3 +37,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - project-code as parameter in project-view
 - Bridgehead state
 - Add URL to publications and other documents
+- User State